refactor: 1차 리팩토링 ts 완료
