+++ package.json
@@ -57,7 +57,11 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@types/date-fns": "^2.6.0",
+    "@types/lodash": "^4.14.197",
+    "@types/lodash.range": "^3.2.7",
     "@types/react": "^18.2.0",
+    "@types/react-datepicker": "^4.15.0",
     "@types/react-dom": "^18.2.0",
     "json-server": "^0.17.3",
     "typescript": "^5.1.6"
