+++ src/components/crewpost/CrewTypes.ts
@@ -1,0 +1,12 @@
+export interface Crew {
+ title: string;
+ content: string;
+ crewName: string;
+ location: string;
+ usersLocation: string;
+ exerciseDate: string;
+ exerciseKind: string;
+ totalNumber: number;
+ images: File[];
+ time: string;
+}
